backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.9.0
python-multipart==0.0.12

# Garmin Connect (Community Library)
garminconnect==0.2.38
# garth is required by garminconnect; pinned to 0.5.21 for compatibility
garth==0.5.21

# CORS
python-dotenv==1.0.1

# HTTP Client (for Strava API)
httpx==0.27.2
